✨ feat(fuzz): add tarn bytecode VM fuzz target
Adds a `tarn` fuzz target so the arbitrary-script generator already used by
`interpreter` can also be run against the tarn bytecode VM instead of the
tree-walking evaluator, gated behind mq-lang's `tarn` feature.

Shared script-generation and eval logic moves into fuzz/src/lib.rs so both
targets stay in sync. `just test-fuzz-tarn` runs the new target.
